+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 ## Entries
 
+## v1.2.3
+
+- Fix for polygon click-through causing incorrect fill color, Issues #174 #154
+- Support global value mappings in Overrides #164, Issue #159
+- Compute X and Y mouse positions in mouseover #166, Issue #163
+- Increase polystat-panel-tooltip z-index #165, Issue #130
+- Fix for composite threshold evalation using global thresholds #174, Issues #158 and #173
+- Fix for default warning color not rendering corectly #179
+- Feature: Tooltip can now be disabled (on by default), PR 158
+
+## v1.2.2
+
+- Fix for gradient threshold color issue #126
+
 ## v1.2.1
 
 - Update for v6 event compatibility